Move RecognitionException to ncore; drop getURIFromTokenSource delegate (#3867)
Follow-on cleanup enabled by moving the source-location classes to
key.ncore:

  * RecognitionException is a generic, language-independent ANTLR-style
    error (implements HasLocation, uses only Position/Location/IntStream).
    Its only key.core dependency was MiscTools.getURIFromTokenSource; with
    that relocated to SourceNames it moves to key.ncore
    (org.key_project.util.parsing.RecognitionException).
  * MiscTools.getURIFromTokenSource had become a pass-through to
    SourceNames; remove it and let the remaining callers use SourceNames
    directly.
  * ExceptionTools.getLocation(RecognitionException) (and its only helper
    parseFileName) is now redundant: RecognitionException implements
    HasLocation, so the generic `instanceof HasLocation` branch covers it.
    Removed, dropping another special-case and some dead imports.

Created with AI tooling support
