+++ key.core/src/main/java/de/uka/ilkd/key/control/instantiation_model/TacletAssumesModel.java
@@ -18,7 +18,6 @@ import de.uka.ilkd.key.proof.MissingInstantiationException;
 import de.uka.ilkd.key.proof.SVInstantiationParserException;
 import de.uka.ilkd.key.proof.io.ProofSaver;
 import de.uka.ilkd.key.rule.TacletApp;
-import de.uka.ilkd.key.util.RecognitionException;
 
 import org.key_project.logic.Term;
 import org.key_project.prover.rules.instantiation.AssumesFormulaInstDirect;
@@ -26,6 +25,7 @@ import org.key_project.prover.rules.instantiation.AssumesFormulaInstantiation;
 import org.key_project.prover.sequent.SequentFormula;
 import org.key_project.util.collection.ImmutableList;
 import org.key_project.util.parsing.Position;
+import org.key_project.util.parsing.RecognitionException;
 
 public class TacletAssumesModel extends DefaultComboBoxModel<AssumesFormulaInstantiation> {
