+++ key.core/src/main/java/de/uka/ilkd/key/util/ExceptionTools.java
@@ -5,8 +5,6 @@ package de.uka.ilkd.key.util;
 
 import java.io.File;
 import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -24,6 +22,7 @@ import de.uka.ilkd.key.util.parsing.SyntaxErrorReporter;
 import org.key_project.util.parsing.HasLocation;
 import org.key_project.util.parsing.Location;
 import org.key_project.util.parsing.Position;
+import org.key_project.util.parsing.SourceNames;
 
 import org.antlr.v4.runtime.*;
 import org.antlr.v4.runtime.misc.IntervalSet;
@@ -341,7 +340,7 @@ public final class ExceptionTools {
         if (ip == null) {
             return Location.fromToken(offending);
         }
-        return new Location(MiscTools.getURIFromTokenSource(offending.getTokenSource()), ip);
+        return new Location(SourceNames.getURIFromTokenSource(offending.getTokenSource()), ip);
     }
 
     /**
@@ -412,13 +411,4 @@ public final class ExceptionTools {
         }
         return false;
     }
-
-    private static URI parseFileName(String filename) throws MalformedURLException {
-        try {
-            return filename == null ? null : MiscTools.parseURL(filename).toURI();
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
 }
