+++ key.core/src/test/java/de/uka/ilkd/key/parser/TestIntLiteralParsing.java
@@ -13,7 +13,8 @@ import de.uka.ilkd.key.logic.op.LocationVariable;
 import de.uka.ilkd.key.speclang.PositionedString;
 import de.uka.ilkd.key.speclang.njml.JmlIO;
 import de.uka.ilkd.key.speclang.njml.SpecMathMode;
-import de.uka.ilkd.key.util.RecognitionException;
+
+import org.key_project.util.parsing.RecognitionException;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DynamicTest;
